Refresh entity on each request; merge EntityHome classes
