+++ zanata-war/src/main/java/org/zanata/action/ProjectHome.java
@@ -94,8 +94,6 @@ import static javax.faces.application.FacesMessage.SEVERITY_ERROR;
 
 @Named("projectHome")
 @Slf4j
-//@GroupedConversationScoped
-//@ConversationGroup(ProjectSlug.class)
 @ViewScoped
 public class ProjectHome extends SlugHome<HProject> implements
     HasLanguageSettings {
@@ -283,9 +281,10 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     public void createNew() {
         clearSlugs();
-        identity.checkPermission(getInstance(), "insert");
-        getInstance().setDefaultProjectType(ProjectType.File);
-        selectedProjectType = getInstance().getDefaultProjectType().name();
+        HProject instance = getInstance();
+        identity.checkPermission(instance, "insert");
+        instance.setDefaultProjectType(ProjectType.File);
+        selectedProjectType = ProjectType.File.name();
         enteredLocaleAliases.putAll(getLocaleAliases());
         // force get so it will create and populate the hashmap
         getSelectedEnabledLocales();
@@ -364,7 +363,7 @@ public class ProjectHome extends SlugHome<HProject> implements
      */
     @Transactional
     public void updateToEnteredLocaleAlias(LocaleId localeId) {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         String enteredAlias = enteredLocaleAliases.get(localeId);
         setLocaleAlias(localeId, enteredAlias);
     }
@@ -425,7 +424,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void removeSelectedLocaleAliases() {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         List<LocaleId> removed = new ArrayList<>();
         for (Map.Entry<LocaleId, Boolean> entry :
                 getSelectedEnabledLocales().entrySet()) {
@@ -441,7 +440,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void removeAllLocaleAliases() {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         List<LocaleId> removed = new ArrayList<>();
         List<LocaleId> aliasedLocales =
                 new ArrayList<>(getLocaleAliases().keySet());
@@ -476,7 +475,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void disableLocale(HLocale locale) {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         disableLocaleSilently(locale);
         facesMessages.addGlobal(FacesMessage.SEVERITY_INFO,
                 msgs.format("jsf.languageSettings.LanguageDisabled",
@@ -486,7 +485,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void disableSelectedLocales() {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         List<LocaleId> removedLocales = new ArrayList<>();
         for (Map.Entry<LocaleId, Boolean> entry :
                 getSelectedEnabledLocales().entrySet()) {
@@ -534,7 +533,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void enableLocale(HLocale locale) {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         enableLocaleSilently(locale);
         LocaleId localeId = locale.getLocaleId();
         facesMessages.addGlobal(FacesMessage.SEVERITY_INFO,
@@ -543,7 +542,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void enableSelectedLocales() {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         List<LocaleId> addedLocales = new ArrayList<>();
         for (Map.Entry<LocaleId, Boolean> entry : selectedDisabledLocales
                 .entrySet()) {
@@ -590,7 +589,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void useDefaultLocales() {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         setOverrideLocales(false);
         removeAliasesForDisabledLocales();
         refreshDisabledLocales();
@@ -652,7 +651,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void setRestrictedByRole(String key, boolean checked) {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         getInstance().setRestrictedByRoles(checked);
         update();
     }
@@ -720,7 +719,6 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     // @Begin(join = true) /* TODO [CDI] commented out begin conversation. Verify it still works properly */
     public void initialize() {
-        initInstance();
         validateSuppliedId();
         if (getInstance().getDefaultCopyTransOpts() != null) {
             copyTransOptionsModel.setInstance(getInstance()
@@ -770,7 +768,7 @@ public class ProjectHome extends SlugHome<HProject> implements
     @Override
     @Transactional
     public String update() {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         // getInputSlugValue() can be null
         if (!getSlug().equals(getInputSlugValue()) && !validateSlug(getInputSlugValue(), "slug")) {
             return null;
@@ -859,7 +857,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public String removeMaintainer(HPerson person) {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         if (getInstanceMaintainers().size() <= 1) {
             facesMessages.addGlobal(FacesMessage.SEVERITY_INFO,
                 msgs.get("jsf.project.NeedAtLeastOneMaintainer"));
@@ -880,7 +878,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void updateRoles(String roleName, boolean isRestricted) {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         getInstance().getAllowedRoles().clear();
         if (getInstance().isRestrictedByRoles()) {
             getRoleRestrictions().put(roleName, isRestricted);
@@ -900,7 +898,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void updateStatus(char initial) {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         getInstance().setStatus(EntityStatus.valueOf(initial));
         if (getInstance().getStatus() == EntityStatus.READONLY) {
             for (HProjectIteration version : getInstance()
@@ -1023,7 +1021,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void updateValidationOption(String name, String state) {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         ValidationId validationId = ValidationId.valueOf(name);
 
         for (Map.Entry<ValidationId, ValidationAction> entry : getValidations()
@@ -1054,7 +1052,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void addWebHook(String url, String secret) {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         if (isValidUrl(url)) {
             secret = StringUtils.isBlank(secret) ? null : secret;
             WebHook webHook = new WebHook(this.getInstance(), url, secret);
@@ -1067,7 +1065,7 @@ public class ProjectHome extends SlugHome<HProject> implements
 
     @Transactional
     public void removeWebHook(Long webhookId) {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         WebHook webHook = webHookDAO.findById(webhookId);
         if (webHook != null) {
             getInstance().getWebHooks().remove(webHook);
@@ -1122,7 +1120,7 @@ public class ProjectHome extends SlugHome<HProject> implements
      */
     @Transactional
     public void updateAboutPage() {
-        identity.checkPermission(instance, "update");
+        identity.checkPermission(getInstance(), "update");
         String status = update();
         if ("updated".equals(status)) {
             facesMessages.addGlobal(FacesMessage.SEVERITY_INFO,
