+++ zanata-war/src/main/java/org/zanata/action/SlugHome.java
@@ -23,6 +23,7 @@ package org.zanata.action;
 import org.apache.deltaspike.core.spi.scope.conversation.GroupedConversationManager;
 import org.hibernate.Session;
 import org.hibernate.criterion.NaturalIdentifier;
+import org.zanata.model.ModelEntityBase;
 import org.zanata.seam.framework.EntityHome;
 
 import javax.inject.Inject;
@@ -32,7 +33,7 @@ import javax.inject.Inject;
  *
  * @author asgeirf
  */
-public abstract class SlugHome<E> extends EntityHome<E> {
+public abstract class SlugHome<E extends ModelEntityBase> extends EntityHome<E> {
 
     private static final long serialVersionUID = 1L;
