+++ zanata-war/src/main/java/org/zanata/seam/framework/EntityHome.java
@@ -2,25 +2,50 @@
 
 package org.zanata.seam.framework;
 
+import javax.annotation.Nonnull;
+import javax.annotation.PostConstruct;
+import javax.faces.application.FacesMessage;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
+import javax.persistence.EntityNotFoundException;
+import javax.transaction.Status;
 import javax.transaction.SystemException;
 import javax.transaction.UserTransaction;
 
 import org.apache.deltaspike.jpa.api.transaction.Transactional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.zanata.model.ModelEntityBase;
+import org.zanata.ui.faces.FacesMessages;
 import org.zanata.util.ServiceLocator;
+
+import java.io.Serializable;
+import java.lang.reflect.ParameterizedType;
+import java.lang.reflect.Type;
+import java.lang.reflect.TypeVariable;
 
 import static javax.transaction.Status.STATUS_ACTIVE;
 import static javax.transaction.Status.STATUS_MARKED_ROLLBACK;
 
 /**
+ * Base class for components which provide persistence operations to a JPA-
+ * managed entity instance. This class may be reused by either configuration
+ * or extension, and may be bound directly to a view, or accessed by some
+ * intermediate Seam component.
+ *
  * Base class for Home objects of JPA entities.
  *
  * @author Gavin King
  */
-public abstract class EntityHome<E> extends Home<EntityManager, E> {
-    private static final long serialVersionUID = -3140094990727574632L;
+public abstract class EntityHome<E extends ModelEntityBase> implements
+        Serializable {
+    private static final Logger log = LoggerFactory.getLogger(EntityHome.class);
+    private static final long serialVersionUID = 1L;
+
+    private Object id;
+    private E instance;
+    // NonNull: This will be initialised at PostConstruct time
+    private Class<E> entityClass;
 
     @Inject
     private UserTransaction transaction;
@@ -29,16 +54,113 @@ public abstract class EntityHome<E> extends Home<EntityManager, E> {
     private EntityManager entityManager;
 
     /**
-     * Run on {@link EntityHome} instantiation. <br />
-     * Validates that an {@link EntityManager} is available.
-     *
-     * @see Home#create()
+     * Run on {@link Home} instantiation to check the Home component is in a
+     * valid state. <br />
+     * Validates that the class of the entity to be managed has been specified.
      */
-    @Override
-    public void create() {
-        super.create();
-        if (getEntityManager() == null) {
-            throw new IllegalStateException("entityManager is null");
+    @PostConstruct
+    public void postConstruct() {
+        if (getEntityClass() == null) {
+            throw new IllegalStateException("entityClass is null");
+        }
+    }
+
+    /**
+     * Get the id of the object being managed.
+     */
+    public Object getId() {
+        return id;
+    }
+
+    /**
+     * Set/change the entity being managed by id.
+     *
+     */
+    public void setId(Object id) {
+        setInstance(null);
+        this.id = id;
+    }
+
+    /**
+     * Set the id of entity being managed. <br />
+     * Does not alter the instance so used if the id of the managed object is
+     * changed.
+     *
+     */
+    protected void assignId(Object id) {
+        this.id = id;
+    }
+
+    /**
+     * Returns true if the id of the object managed is known (eg from a
+     * servlet request parameter). Note that this might be true for one
+     * request but not for another, during the ViewScoped lifetime of the
+     * EntityHome object (so the instance should remember its id).
+     */
+    public boolean isIdDefined() {
+        return getId() != null && !"".equals(getId());
+    }
+
+    /**
+     * Get the managed entity, using the id from {@link #getId()} to load, it
+     * from the Persistence Context, refreshing it if disconnected,
+     * or creating a new instance if the id is not
+     * defined. Because our EntityManagers have shorter scopes than EntityHome, we
+     * reload the entity on each request to ensure it is attached.
+     *
+     * @see #getId()
+     */
+    @Transactional
+    public @Nonnull E getInstance() {
+        joinTransaction();
+        if (instance == null) {
+            if (isIdDefined()) {
+                this.instance = loadInstanceByExternalId();
+            } else {
+                this.instance = createInstance();
+            }
+        } else if (instance.getId() != null) {
+            if (!entityManager.contains(instance)) {
+                E result = entityManager.find(getEntityClass(), instance.getId());
+                if (result == null) {
+                    handleNotFound();
+                } else {
+                    this.instance = result;
+                }
+            }
+        }
+        return instance;
+    }
+
+    /**
+     * Set/change the entity being managed.
+     */
+    protected void setInstance(E instance) {
+        this.instance = instance;
+    }
+
+    private E loadInstanceByExternalId() {
+        if (!isTransactionMarkedRollback()) {
+            // we cache the instance so that it does not "disappear"
+            // after remove() is called on the instance
+            // is this really a Good Idea??
+            return find();
+        } else {
+            throw new RuntimeException("transaction rolled back");
+        }
+    }
+
+    /**
+     * Create a new instance of the entity. <br />
+     * Utility method called by {@link #initInstance()} to create a new instance
+     * of the entity.
+     */
+    protected E createInstance() {
+        assert entityClass != null;
+        try {
+            return entityClass.newInstance();
+        } catch (Exception e) {
+            throw new RuntimeException(e);
         }
     }
 
@@ -47,8 +169,7 @@ public abstract class EntityHome<E> extends Home<EntityManager, E> {
      */
     @Transactional
     public boolean isManaged() {
-        return getInstance() != null &&
-                getEntityManager().contains(getInstance());
+        return getEntityManager().contains(getInstance());
     }
 
     /**
@@ -85,14 +206,8 @@ public abstract class EntityHome<E> extends Home<EntityManager, E> {
         getEntityManager().persist(getInstance());
         getEntityManager().flush();
         E instance = getInstance();
-        if (instance instanceof ModelEntityBase) {
-            Long id = ((ModelEntityBase) instance).getId();
-            assignId(id);
-        } else {
-            // this should not happen. We don't have an EntityHome<E> where E is
-            // not a subclass of ModelEntityBase
-            throw new RuntimeException("can not get id from entity");
-        }
+        Long id = instance.getId();
+        assignId(id);
 
         createdMessage();
         return "persisted";
@@ -115,13 +230,62 @@ public abstract class EntityHome<E> extends Home<EntityManager, E> {
     }
 
     /**
-     * Implementation of {@link Home#find() find()} for JPA
+     * Add a {@link javax.faces.application.FacesMessage} and log a message when
+     * the entity instance is updated.
+     * <p/>
+     * Utility method to add a {@link javax.faces.application.FacesMessage} from
+     * the Seam managed resource bundle and log the entity when the
+     * managed entity is updated.
      *
-     * @see Home#find()
+     */
+    protected void updatedMessage() {
+        log.debug("updated entity {} {}", getEntityClass().getName(), getId());
+        getFacesMessages().addGlobal(
+                FacesMessage.SEVERITY_INFO, "Successfully updated");
+    }
+
+    private FacesMessages getFacesMessages() {
+        return ServiceLocator.instance().getInstance(FacesMessages.class);
+    }
+
+    /**
+     * Add a {@link javax.faces.application.FacesMessage} and log a message when
+     * the entity instance is deleted.
+     * <p/>
+     * Utility method to add a {@link javax.faces.application.FacesMessage} from
+     * the Seam managed resource bundle and log the entity when the
+     * managed entity is deleted.
+     *
+     */
+    protected void deletedMessage() {
+        log.debug("deleted entity {} {}", getEntityClass().getName(), getId());
+        getFacesMessages().addGlobal(
+                FacesMessage.SEVERITY_INFO, "Successfully deleted");
+
+    }
+
+    /**
+     * Add a {@link javax.faces.application.FacesMessage} and log a message when
+     * the entity instance is created.
+     * <p/>
+     * Utility method to add a {@link javax.faces.application.FacesMessage} from
+     * the Seam managed resource bundle and log the entity when the
+     * managed entity is updated.
+     *
+     */
+    protected void createdMessage() {
+        log.debug("created entity {} {}", getEntityClass().getName(), getId());
+        getFacesMessages().addGlobal(
+                FacesMessage.SEVERITY_INFO, "Successfully created");
+
+    }
+
+    /**
+     * Hook method called by {@link #initInstance()} to allow the implementation
+     * to load the entity from the Persistence Context.
      */
     @Transactional
-    @Override
-    public E find() {
+    public @Nonnull E find() {
         if (getEntityManager().isOpen()) {
             E result = loadInstance();
             if (result == null) {
@@ -129,8 +293,17 @@ public abstract class EntityHome<E> extends Home<EntityManager, E> {
             }
             return result;
         } else {
-            return null;
+            throw new RuntimeException("EntityManager is closed");
         }
+    }
+
+    /**
+     * Utility method called by the framework when no entity is found in the
+     * Persistence Context.
+     */
+    protected E handleNotFound() {
+        throw new EntityNotFoundException(String.format(
+                "entity not found: %s#%s", getEntityClass(), getId()));
     }
 
     /**
@@ -145,11 +318,18 @@ public abstract class EntityHome<E> extends Home<EntityManager, E> {
         return getEntityManager().find(getEntityClass(), getId());
     }
 
+    protected boolean isTransactionMarkedRollback() {
+        try {
+            return getTransaction().getStatus() == Status.STATUS_MARKED_ROLLBACK;
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
     /**
-     * Implementation of {@link Home#joinTransaction() joinTransaction()} for
-     * JPA.
+     * Hook method called to allow the implementation to join the current
+     * transaction when necessary.
      */
-    @Override
     protected void joinTransaction() {
         if (getEntityManager().isOpen()) {
             try {
@@ -170,10 +350,52 @@ public abstract class EntityHome<E> extends Home<EntityManager, E> {
     }
 
     /**
-     * The Seam Managed Persistence Context used by this Home component
+     * The EntityManager used by this Home component
      */
-    public EntityManager getEntityManager() {
+    protected EntityManager getEntityManager() {
         return entityManager;
+    }
+
+    /**
+     * Get the class of the entity being managed. <br />
+     * If not explicitly specified, the generic type of implementation is used.
+     */
+    public Class<E> getEntityClass() {
+        if (entityClass == null) {
+            // CDI will return a proxy instance (so need an extra getSuperClass)
+            Type type = getClass().getSuperclass().getGenericSuperclass();
+            if (type instanceof ParameterizedType) {
+                ParameterizedType paramType = (ParameterizedType) type;
+                if (paramType.getActualTypeArguments().length == 2) {
+                    // likely dealing with -> new
+                    // EntityHome<Person>().getEntityClass()
+                    if (paramType
+                            .getActualTypeArguments()[1] instanceof TypeVariable) {
+                        throw new IllegalArgumentException(
+                                "Could not guess entity class by reflection");
+                    } else {
+                        // likely dealing with -> new Home<EntityManager, Person>()
+                        // { ... }.getEntityClass()
+                        entityClass =
+                                (Class<E>) paramType
+                                        .getActualTypeArguments()[1];
+                    }
+                } else {
+                    // likely dealing with -> new PersonHome().getEntityClass()
+                    // where PersonHome extends EntityHome<Person>
+                    entityClass =
+                            (Class<E>) paramType.getActualTypeArguments()[0];
+                }
+            } else {
+                throw new IllegalArgumentException(
+                        "Could not guess entity class by reflection");
+            }
+        }
+        return entityClass;
+    }
+
+    public void setEntityClass(Class<E> entityClass) {
+        this.entityClass = entityClass;
     }
 
 }
